Works, but still translation from en to pl needs some fixing -> null

+++ src/flats/entities/flats-gpt.entity.ts
@@ -154,7 +154,7 @@ export class FlatsGPT {
   outbuildingSummary: string | null;
 
   @Column({
-    type: 'tinyint',
+    type: 'int',
     default: null,
   })
   rentRating: number | null;
